move start quiz to top
